Fixed horizontal containment radius limit in surface position (#23)
Removed wrong cast to byte

+++ src/main/java/de/serosystems/lib1090/msgs/adsb/SurfacePositionV2Msg.java
@@ -81,7 +81,7 @@ public class SurfacePositionV2Msg extends SurfacePositionV1Msg implements Serial
 			case 5: return 7.5;
 			case 6: return 25;
 			case 7:
-				return (byte) (hasNICSupplementA() ? 75 : 185.2);
+				return hasNICSupplementA() ? 75 : 185.2;
 			case 8:
 				if (hasNICSupplementC() && hasNICSupplementA())
 					return 370.4;
